chore(event_handler): move cookies to shared (cicular dependency)
