+++ aws_lambda_powertools/shared/cookies.py
